requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cu130
# Core ML libraries
torch==2.10.0+cu130
torchaudio==2.10.0+cu130
torchvision==0.25.0+cu130


# Generative AI & Hugging Face
diffusers==0.36.0
transformers==5.0.0
accelerate==1.12.0
scipy==1.17.0
ftfy==6.3.1
python-dotenv==1.2.1

# Utilities
Pillow==12.0.0
numpy==2.3.5

# Gradio UI
gradio==6.5.1